note on what's needed for correct order of operations on accreted views
accreted views are views that sometimes have a key that is overwritten
as newer data comes in. to ensure that we're actually overwriting with
the newest data, we have to have knowledge that that is in fact true. we
can accomplish that by either changing the key layout so that we can
range & always get the latest message, or we can compare incoming
messages to the timestamps of the latest written message for each
possibl key