adds a sentinel to channel membership view to accomodate channel list responses
regular key scheme is: `<channel>!<pubkeyHex>: 0|1`

sentinel key scheme is: `<channel>!<'x'.repeat(64)>: 0`

the reason for this is to accomodate channel list responses. this
response type acts as a way for people to discover channels without
needing to sync history they don't care about. it does this by returning
a list of channel names. however, this list of channel names needs to be
indexed by some view. in this commit we introduce one way to keep track
of these channels, which is by using a sentinel in place of an actual
publicKey hex (as we have none).

using the sentinel allows us to distinguish from actual user membership
activity while also being able to keep track of which channels we know
about without adding another view