reindex many channels on delete, add evt manager
this commit makes a substantial change to the way deletes are handled in
that the behaviour for handling post/info is now more correct. the code
now takes into account that deletion of a post/info does not a have
channel property. furthermore, as what is being changed affects a user,
which may belong to many channels, and not only a single channel we now
propagate that information to the many channels that may be affected.

finally, we now emit an event to reflect that the previous latest
message in a particular view (be it channel topic, channel membership,
or user name) has been deleted and replaced with a new hash. this new
hash is what is emitted, bringing the behaviour more in line with the
channel state request's intende behaviour (per spec rev 2023-04)

finally a manager class, EventsManager, is introduced to simplify
correct handling, and debugging, of all the different event listeners
that will be part of this library.