test links
this commit adds unit tests for exercising the links functionality. as
part of implementing the tests we've caught some smol straggling bugs,
which is why there are additions to index.js + links.js

as of this time, the basic links & heads tracking functionality of cable when authoring posts &
ingesting posts coming in from another peer has been confirmed!