index.js: clean up debug logging output
the previous debug logging was useful but a bit haphazard and too
verbose. this pass places the request id (a useful piece of information
when tracking down behaviour across many clients) in a consistent spot
for all of the relevant logs. verbosity has also been tamed even if it
is probably not perfect :)

as it is, this was a useful change when debugging curious behaviour and
i remain glad of the decision to go all in on and retaqin the granular
logging built up in cable-core and cable-client