handle post response edge case when connected to multiple peers
see https://github.com/cabal-club/cable/issues/25 for context, but in
brief:

there's currently a bit of behaviour i'm uncertain how to handle in
regards to the current structure of keeping track of "alive" requests,
what the spec asks for in relation to handling post responses, and
handling "concluding post responses". basically, if we stop tracking a
request id as soon as we receive a concluding post response (post
response with empty field `posts`) then we may shut out important
information that we have asked for.

ran into this when running 3 simultaneous sessions and noticing that the
third session was often not appearing in the nicks sidebar of cable-cli,
despite having verified data flowing across all connections.

it turned out that the second peer managed to 'conclude' the post
request before the third peer managed to send off their response with
the sought-after posts. this patch solves this particular scenario by
simply not responding to a post request at all if a peer has none of the
requested data (i.e. not sending a concluding post response posts=[]).
but related scenarios may still occur and so the issue more or less
persists until we reach consensus in the discussion of the issue posted
at the top of this commit message