use lru cache to mitigate emitting duplicate events
in some cases a hash may be handled multipled times. it would be best to
find the root cause in each and every of those instances, of course. but
for the time between detecting the problem and finding the root cause of such a problem,
the lru cache helps with keeping the experience in clients tidy by only
emitting as many events as needed