add block
this commit introduced an alpha version of blocking. it is a basic
version of the functionality but which proves to work in cable-cli

dropping is not implemented (as it is a larger batch of work and not
implemented for other functionality either).

a user that is blocked will no longer have its future posts stored
(their old posts will continue to be stored, as that is part of drop=1
functionality).

if a node communicates with a user that has blocked them then
the node will not store nor see the blocker's posts.